HELP PLEASE lasso & ridge predictions
I stayed up really late trying to get my prediction code to work for lasso and ridge. I can't seem to get it to work for the life of me. I'm unable to write my analysis of this portion until it's figured out. Any advice would be great... Thanks so much.